# internal/docs/swagger.yaml
basePath: /api
definitions:
  httpsuite.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status_code:
        type: integer
    type: object
  models.ACL:
    properties:
      group:
        type: string
      permissions:
        items:
          enum:
          - read
          - write
          - admin
          type: string
        type: array
      vhost_ids:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    name: Norsk helsenett SF
    url: https://github.com/NorskHelsenett/unimq-dashboard
  description: API for RabbitMQ Dashboard application.
  title: RabbitMQ Dashboard API
  version: "1.0"
paths:
  /v1/acls:
    get:
      description: Get all vhost access control lists. Requires membership in an admin
        group.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ACL'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      summary: Get ACLs
      tags:
      - ACLs
    put:
      consumes:
      - application/json
      description: Create or replace a group's vhost access control list. Requires
        membership in an admin group.
      parameters:
      - description: ACL definition
        in: body
        name: acl
        required: true
        schema:
          $ref: '#/definitions/models.ACL'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ACL'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      summary: Create or update an ACL
      tags:
      - ACLs
  /v1/acls/{group}:
    delete:
      description: Delete a group's vhost access control list. Requires membership
        in an admin group.
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ACL deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      summary: Delete an ACL
      tags:
      - ACLs
    get:
      description: Get the access control list for a group. Requires membership in
        an admin group.
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ACL'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      summary: Get an ACL
      tags:
      - ACLs
...
securityDefinitions:
  OAuth2:
    authorizationUrl: http://localhost:5556/dex/auth
    flow: accessCode
    scopes:
      audience:server:client_id:unimq-dashboard: Issue a token whose audience the
        dashboard API accepts
      email: User email address
      groups: Group membership used for authorization
      openid: OpenID Connect identity
      profile: User profile
    tokenUrl: http://localhost:5556/dex/token
    type: oauth2
  bearer:
    description: '"JWT token for authentication, obtained from the Dex OIDC provider,
      formatted as ''Bearer {token}'' in the Authorization header"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"